refactored project card comp util functions
